+++ fabric/src/main/java/cc/cassian/item_descriptions/client/fabric/ItemDescriptionsFabricClient.java
@@ -1,8 +1,19 @@
 package cc.cassian.item_descriptions.client.fabric;
 
+import cc.cassian.item_descriptions.client.DescriptionKey;
 import cc.cassian.item_descriptions.client.ModClient;
+import cc.cassian.item_descriptions.client.config.ModConfig;
+import cc.cassian.item_descriptions.client.helpers.ModHelpers;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.item.v1.ItemTooltipCallback;
+import net.minecraft.component.DataComponentTypes;
+import net.minecraft.component.type.ItemEnchantmentsComponent;
+import net.minecraft.text.Text;
+import net.minecraft.text.TextContent;
+import net.minecraft.text.TranslatableTextContent;
+
+import java.awt.*;
+import java.util.HashSet;
 
 import static cc.cassian.item_descriptions.client.helpers.ModHelpers.*;
 
@@ -20,7 +31,7 @@ public final class ItemDescriptionsFabricClient implements ClientModInitializer 
         //?} else
         /*ItemTooltipCallback.EVENT.register((stack, context, lines) -> {*/
             //Only show tooltip if key is pressed or "always on" is enabled.
-            createItemDescription(stack, lines);
+            createDescriptionsFromItemStack(stack, lines);
         });
     }
 }